chore(ui/dataLoaders): Move the dataLoaders sidebar, utils, and constants to dataLoaders dir
