+++ ui/src/dataLoaders/components/side_bar/SideBar.scss
