+++ ui/src/dataLoaders/components/side_bar/SideBar.test.tsx
@@ -3,10 +3,10 @@ import React from 'react'
 import {shallow} from 'enzyme'
 
 // Components
-import SideBar from 'src/onboarding/components/side_bar/SideBar'
+import SideBar from 'src/dataLoaders/components/side_bar/SideBar'
 
 // Types
-import {SideBarTabStatus as TabStatus} from 'src/onboarding/components/side_bar/SideBar'
+import {SideBarTabStatus as TabStatus} from 'src/dataLoaders/components/side_bar/SideBar'
 
 // Constants
 import {ComponentColor} from 'src/clockface'
