+++ ui/src/dataLoaders/graphics/IconCSV.tsx
