+++ ui/src/dataLoaders/graphics/LogoApache.tsx
