+++ ui/src/dataLoaders/graphics/LogoConsul.tsx
