+++ ui/src/dataLoaders/graphics/LogoCpu.tsx
