+++ ui/src/dataLoaders/graphics/LogoElastic.tsx
