+++ ui/src/dataLoaders/graphics/LogoEtcd.tsx
