+++ ui/src/dataLoaders/graphics/LogoMesos.tsx
