+++ ui/src/dataLoaders/graphics/LogoMysql.tsx
