+++ ui/src/dataLoaders/graphics/LogoPrometheus.tsx
