+++ ui/src/dataLoaders/graphics/index.ts
@@ -1,0 +1,37 @@
+import LogoApache from 'src/dataLoaders/graphics/LogoApache'
+import LogoConsul from 'src/dataLoaders/graphics/LogoConsul'
+import LogoCpu from 'src/dataLoaders/graphics/LogoCpu'
+import LogoDocker from 'src/dataLoaders/graphics/LogoDocker'
+import LogoElastic from 'src/dataLoaders/graphics/LogoElastic'
+import LogoEtcd from 'src/dataLoaders/graphics/LogoEtcd'
+import LogoIis from 'src/dataLoaders/graphics/LogoIis'
+import LogoKubernetes from 'src/dataLoaders/graphics/LogoKubernetes'
+import LogoMesos from 'src/dataLoaders/graphics/LogoMesos'
+import LogoMongodb from 'src/dataLoaders/graphics/LogoMongodb'
+import LogoMysql from 'src/dataLoaders/graphics/LogoMysql'
+import LogoNginx from 'src/dataLoaders/graphics/LogoNginx'
+import LogoRedis from 'src/dataLoaders/graphics/LogoRedis'
+import LogoPrometheus from 'src/dataLoaders/graphics/LogoPrometheus'
+import IconStreaming from 'src/dataLoaders/graphics/IconStreaming'
+import IconCSV from 'src/dataLoaders/graphics/IconCSV'
+import IconLineProtocol from 'src/dataLoaders/graphics/IconLineProtocol'
+
+export {
+  LogoApache,
+  LogoConsul,
+  LogoCpu,
+  LogoDocker,
+  LogoElastic,
+  LogoEtcd,
+  LogoIis,
+  LogoKubernetes,
+  LogoMongodb,
+  LogoMesos,
+  LogoMysql,
+  LogoNginx,
+  LogoRedis,
+  LogoPrometheus,
+  IconStreaming,
+  IconCSV,
+  IconLineProtocol,
+}
