--- ui/src/onboarding/graphics/index.ts
@@ -1,37 +1,0 @@
-import LogoApache from 'src/onboarding/graphics/LogoApache'
-import LogoConsul from 'src/onboarding/graphics/LogoConsul'
-import LogoCpu from 'src/onboarding/graphics/LogoCpu'
-import LogoDocker from 'src/onboarding/graphics/LogoDocker'
-import LogoElastic from 'src/onboarding/graphics/LogoElastic'
-import LogoEtcd from 'src/onboarding/graphics/LogoEtcd'
-import LogoIis from 'src/onboarding/graphics/LogoIis'
-import LogoKubernetes from 'src/onboarding/graphics/LogoKubernetes'
-import LogoMesos from 'src/onboarding/graphics/LogoMesos'
-import LogoMongodb from 'src/onboarding/graphics/LogoMongodb'
-import LogoMysql from 'src/onboarding/graphics/LogoMysql'
-import LogoNginx from 'src/onboarding/graphics/LogoNginx'
-import LogoRedis from 'src/onboarding/graphics/LogoRedis'
-import LogoPrometheus from 'src/onboarding/graphics/LogoPrometheus'
-import IconStreaming from 'src/onboarding/graphics/IconStreaming'
-import IconCSV from 'src/onboarding/graphics/IconCSV'
-import IconLineProtocol from 'src/onboarding/graphics/IconLineProtocol'
-
-export {
-  LogoApache,
-  LogoConsul,
-  LogoCpu,
-  LogoDocker,
-  LogoElastic,
-  LogoEtcd,
-  LogoIis,
-  LogoKubernetes,
-  LogoMongodb,
-  LogoMesos,
-  LogoMysql,
-  LogoNginx,
-  LogoRedis,
-  LogoPrometheus,
-  IconStreaming,
-  IconCSV,
-  IconLineProtocol,
-}
